Adding Navigation Training (Only Question embedding & KGE Embedding Space)
Added nav_training which doesn't use the Answer LLM, instead it relies
on the navigator's embedding position to determine if it has arrived at
the answer or not. Only two embedding spaces are considered: Textual
embeddings for the Question (later will be optional) and the KG
embedding for the actual navigation. The episode is considered solved if
the navigation agent is within an epsilon of error from the answer.