Optional Parameter in KGE train: Saving model if above a certain performance threshold
If the saving_metric parameter is employed, the trained kge model will
be stored if it surpasses a certain threshold. Supported metrics are
MRR, HITS@1, HITS@3, HITS@10. If no valid metric is used, it will
default to an empty string and store the model without any conditions.

Improving Sweep Parameters for KinshipHinton