Change: Making Environment parent class more flexible
Through develpment we ended up with children that disobeyed the outlines
set by the parent. I made it more flexible by simply allowing it to
accept a single parameter of `Any` for both action and observation.
Hopefully we release enough constraints while keeping expressiveness in
what an `action` or `observation` is by letting the user define it
themselves.